requirements.txt
matplotlib==2.0.2
torch==0.2.0.post4
easygui==0.98.1
PySoundFile==0.9.0.post1
librosa==0.5.1
tensorflow==1.3.0
numpy==1.12.1
better_exceptions==0.1.8
scipy==0.19.0
Pillow==4.3.0
soundfile==0.9.0.post1
torchvision==0.1.9
